requirements.txt
web3==6.3.0
py-solc-x==1.1.1
Flask==2.3.3
python-dotenv==1.0.0
